final_project/requirements.txt
scrapy
scrapy-playwright
pymongo
fastapi[standard]
pillow
